common/metadata-intelligence/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sql-agent"
version = "0.1.0"
description = "SQL Agent with Planner/Executor pattern for AWS AgentCore"
authors = [{name = "Developer", email = "dev@example.com"}]
dependencies = [
    "strands-framework>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "streamlit>=1.28.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "boto3>=1.34.0",
    "typing-extensions",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"